requirements.txt
#Specify the versions you use
Django==4.0.4
django-heroku==0.3.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
whitenoise==6.1.0